Refactor fixture `start_platform_api_service` for cleaner organization (#17428)
What is the motivation for this PR?
In PR #16860, we moved the start_platform_api_service fixture to tests/conftest.py, but this approach is neither elegant nor clean. Since it is used by the platform test fixture platform_api_conn, we have relocated it to the same file as platform_api_conn for better organization. Additionally, because pytest does not automatically detect fixtures outside of conftest.py, we now need to explicitly import both start_platform_api_service and platform_api_conn.

How did you do it?
Since the fixture start_platform_api_service is used by the platform test fixture platform_api_conn, we have relocated it to the same file as platform_api_conn for better organization. Additionally, because pytest does not automatically detect fixtures outside of conftest.py, we now need to explicitly import both start_platform_api_service and platform_api_conn.

How did you verify/test it?
Perviously, to avoid the error fixture 'start_platform_api_service' not found, we moved this fixture to tests/conftest.py. This time, I have tested locally, as we explicitly import both start_platform_api_service and platform_api_conn, there is no such error.
