+++ tests/conftest.py
@@ -76,10 +76,6 @@ cache = FactsCache()
 DUTHOSTS_FIXTURE_FAILED_RC = 15
 CUSTOM_MSG_PREFIX = "sonic_custom_msg"
 
-SERVER_FILE = 'platform_api_server.py'
-SERVER_PORT = 8000
-IPTABLES_PREPEND_RULE_CMD = 'iptables -I INPUT 1 -p tcp -m tcp --dport {} -j ACCEPT'.format(SERVER_PORT)
-
 pytest_plugins = ('tests.common.plugins.ptfadapter',
                   'tests.common.plugins.ansible_fixtures',
                   'tests.common.plugins.dut_monitor',
@@ -2863,53 +2859,6 @@ def gnxi_path(ptfhost):
     else:
         gnxipath = "/gnxi/"
     return gnxipath
-
-
-@pytest.fixture(scope='function')
-def start_platform_api_service(duthosts, enum_rand_one_per_hwsku_hostname, localhost, request):
-    duthost = duthosts[enum_rand_one_per_hwsku_hostname]
-    dut_ip = duthost.mgmt_ip
-
-    res = localhost.wait_for(host=dut_ip,
-                             port=SERVER_PORT,
-                             state='started',
-                             delay=1,
-                             timeout=10,
-                             module_ignore_errors=True)
-    if res['failed'] is True:
-
-        res = duthost.command('docker exec -i pmon python3 -c "import sonic_platform"', module_ignore_errors=True)
-        py3_platform_api_available = not res['failed']
-
-        supervisor_conf = [
-            '[program:platform_api_server]',
-            'command=/usr/bin/python{} /opt/platform_api_server.py --port {}'.format('3' if py3_platform_api_available
-                                                                                     else '2', SERVER_PORT),
-            'autostart=True',
-            'autorestart=True',
-            'stdout_logfile=syslog',
-            'stderr_logfile=syslog',
-        ]
-        dest_path = os.path.join(os.sep, 'tmp', 'platform_api_server.conf')
-        pmon_path = os.path.join(os.sep, 'etc', 'supervisor', 'conf.d', 'platform_api_server.conf')
-        duthost.copy(content='\n'.join(supervisor_conf), dest=dest_path)
-        duthost.command('docker cp {} pmon:{}'.format(dest_path, pmon_path))
-
-        src_path = os.path.join('common', 'helpers', 'platform_api', 'scripts', SERVER_FILE)
-        dest_path = os.path.join(os.sep, 'tmp', SERVER_FILE)
-        pmon_path = os.path.join(os.sep, 'opt', SERVER_FILE)
-        duthost.copy(src=src_path, dest=dest_path)
-        duthost.command('docker cp {} pmon:{}'.format(dest_path, pmon_path))
-
-        # Prepend an iptables rule to allow incoming traffic to the HTTP server
-        duthost.command(IPTABLES_PREPEND_RULE_CMD)
-
-        # Reload the supervisor config and Start the HTTP server
-        duthost.command('docker exec -i pmon supervisorctl reread')
-        duthost.command('docker exec -i pmon supervisorctl update')
-
-        res = localhost.wait_for(host=dut_ip, port=SERVER_PORT, state='started', delay=1, timeout=10)
-        assert res['failed'] is False
 
 
 def pytest_collection_modifyitems(config, items):
