+++ tests/platform_tests/api/test_component.py
@@ -5,7 +5,7 @@ from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.platform_api import chassis, component
 from .platform_api_test_base import PlatformApiTestBase
 from tests.common.utilities import skip_release_for_platform
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 
 ###################################################
 # TODO: Remove this after we transition to Python 3
