+++ tests/platform_tests/api/test_fan_drawer.py
@@ -3,7 +3,7 @@ import logging
 import pytest
 
 from tests.common.helpers.platform_api import chassis, fan_drawer
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 
 from .platform_api_test_base import PlatformApiTestBase
