+++ tests/platform_tests/api/test_module.py
@@ -8,7 +8,7 @@ from tests.platform_tests.cli.util import get_skip_mod_list
 from .platform_api_test_base import PlatformApiTestBase
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.dut_utils import ignore_t2_syslog_msgs
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 
 ###################################################
 # TODO: Remove this after we transition to Python 3
