+++ tests/platform_tests/api/test_psu.py
@@ -7,8 +7,7 @@ from tests.common.utilities import skip_release
 from tests.platform_tests.cli.util import get_skip_mod_list
 from .platform_api_test_base import PlatformApiTestBase
 from tests.common.utilities import skip_release_for_platform, wait_until
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
-
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 
 ###################################################
 # TODO: Remove this after we transition to Python 3
