+++ tests/platform_tests/api/test_psu_fans.py
@@ -5,7 +5,7 @@ import os
 import pytest
 
 from tests.common.helpers.platform_api import chassis, psu, psu_fan
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 
 from .platform_api_test_base import PlatformApiTestBase
