+++ tests/platform_tests/api/test_watchdog.py
@@ -6,7 +6,7 @@ import yaml
 import pytest
 from tests.common.helpers.platform_api import watchdog
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.platform.device_utils import platform_api_conn    # noqa F401
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service    # noqa F401
 from .platform_api_test_base import PlatformApiTestBase
 
 from collections import OrderedDict
