+++ tests/smartswitch/common/device_utils_dpu.py
@@ -4,7 +4,7 @@ Helper script for DPU  operations
 import logging
 import pytest
 import re
-from tests.common.platform.device_utils import platform_api_conn  # noqa: F401,F403
+from tests.common.platform.device_utils import platform_api_conn, start_platform_api_service  # noqa: F401,F403
 from tests.common.helpers.platform_api import chassis, module
 from tests.common.utilities import wait_until
 from tests.common.helpers.assertions import pytest_assert
